refactor: move VariableRow into separate file
