+++ src/components/variablesContent.tsx
@@ -1,29 +1,13 @@
 'use client';
 
-import {
-  ButtonWithTooltip,
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormMessage,
-  Input,
-} from '@components';
+import type { Variables } from '@types';
+import { ButtonWithTooltip, Form, VariableRow } from '@components';
 import { useTranslations } from 'next-intl';
-import { Key, useEffect, useState } from 'react';
-import {
-  useFieldArray,
-  UseFieldArrayRemove,
-  useForm,
-  UseFormProps,
-  UseFormReturn,
-} from 'react-hook-form';
-import { Copy, X, SquarePlus } from 'lucide-react';
-import { z } from 'zod';
+import { useEffect, useState } from 'react';
+import { useFieldArray, useForm } from 'react-hook-form';
+import { SquarePlus } from 'lucide-react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { VariablesSchema } from '@entities';
-
-type Variables = z.infer<typeof VariablesSchema>;
 
 type Props = {
   userEmail: string;
@@ -113,63 +97,6 @@ export function VariablesContent({ userEmail }: Props) {
         <div className="absolute bottom-10 right-15 p-3 bg-destructive/15 border border-destructive/50 rounded-md w-auto">
           <p className="text-sm text-destructive">{isError}</p>
         </div>
-      )}
-    </div>
-  );
-}
-
-type VariableRowProps = {
-  index: number;
-  form: UseFormReturn<Variables>;
-  remove: (index: number) => void;
-};
-
-function VariableRow({ index, form, remove }: VariableRowProps) {
-  const t = useTranslations('variables');
-
-  const keyPath = `variables.${index}.key` as const;
-  const valuePath = `variables.${index}.value` as const;
-
-  const copyHandler = () => {
-    const keyValue = form.getValues(keyPath)?.trim();
-    if (keyValue) {
-      navigator.clipboard.writeText(`{{${keyValue}}}`);
-    }
-  };
-
-  return (
-    <div className="flex gap-2 w-full items-start">
-      <FormField
-        control={form.control}
-        name={keyPath}
-        render={({ field }) => (
-          <FormItem className="gap-y-0">
-            <FormControl>
-              <Input placeholder={`{{${t('key')}}}`} {...field} />
-            </FormControl>
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-
-      {ButtonWithTooltip(copyHandler, <Copy size={20} />, t('copyTooltip'))}
-
-      <FormField
-        control={form.control}
-        name={valuePath}
-        render={({ field }) => (
-          <FormItem className="flex-1">
-            <FormControl>
-              <Input placeholder={t('value')} {...field} />
-            </FormControl>
-          </FormItem>
-        )}
-      />
-
-      {ButtonWithTooltip(
-        () => remove(index),
-        <X size={20} />,
-        t('removeTooltip')
       )}
     </div>
   );
