+++ src/types/index.ts
@@ -1,1 +1,1 @@
-export type { Method } from './types';
+export type { Method, Variables } from './types';
